fix: retrofit관련부분 data 파일로 이동
