+++ data/src/main/java/com/bestapp/rice/data/notification/PushMsgJson.kt
@@ -1,4 +1,7 @@
-package com.bestapp.rice.ui.notification.kakaonoti.notientity
+package com.bestapp.rice.data.notification
+
+import retrofit2.http.Field
+
 
 data class PushMsgJson(
     val collapse: String,
