+++ data/src/main/java/com/bestapp/rice/data/notification/RegisterTokenEntity.kt
@@ -1,6 +1,4 @@
-package com.bestapp.rice.ui.notification.kakaonoti.notientity
-
-
+package com.bestapp.rice.data.notification
 
 data class RegisterTokenEntity(
     //@SerializedName("NONE")
